fix:properties issues with expiration, hibernate auto validate, docker-compose version, loanServiceImplementation for consistency fields, securityConfig changes and Loan entity consistency field
