+++ src/main/resources/application.properties
@@ -11,11 +11,13 @@ spring.datasource.driver-class-name=org.postgresql.Driver
 
 # Configuración de JPA/Hibernate
 # Esta línea le dice a Hibernate que cree o actualice las tablas
-# en la base de datos automáticamente basándose en nuestras clases @Entity.
-spring.jpa.hibernate.ddl-auto=update
+# en la base de datos automáticamente basándose en nuestras clases @Entity. 
+# Cambiamos a 'validate' una vez que el esquema es estable para evitar cambios accidentales.
+spring.jpa.hibernate.ddl-auto=validate
 
 # (Opcional pero muy útil para desarrollo) Muestra en la consola las consultas SQL que Hibernate ejecuta.
 spring.jpa.show-sql=true
 # Para una clave real, usa un generador online para una cadena Base64 de 256 bits o más.
 application.security.jwt.secret-key=c2VjcmV0LWtleS1mb3ItYm5wbC10ZXN0LXdpdGgtZW5vdWdoLWJ5dGVzLXRvLWJlLXNlY3VyZQ==
-application.security.jwt.expiration=60 # en minutos
+# en minutos
+application.security.jwt.expiration=60
